Se trabajó en la implementación de imegenes y relaciones entre entidades

+++ src/main/java/com/example/manosalaobrabackend/service/ComentarioService.java
@@ -1,12 +1,14 @@
 package com.example.manosalaobrabackend.service;
 
-import com.example.manosalaobrabackend.model.Comentario;
-import com.example.manosalaobrabackend.repository.ComentarioRepository;
+import java.time.LocalDateTime;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
+import com.example.manosalaobrabackend.exceptions.ComentarioNotFoundException;
+import com.example.manosalaobrabackend.model.Comentario;
+import com.example.manosalaobrabackend.repository.ComentarioRepository;
 
 @Service
 public class ComentarioService {
@@ -18,19 +20,36 @@ public class ComentarioService {
         return comentarioRepository.findAll();
     }
 
-    public Optional<Comentario> getComentarioById(Integer id) {
-        return comentarioRepository.findById(id);
-    }
-
-    public Comentario saveComentario(Comentario comentario) {
+    public Comentario postComentario(Comentario comentario) {
+    	return comentarioRepository.save(comentario);
+    	} 
+    
+  //Método delete User
+  	public void deleteComentario(Long id) {
+  		if(comentarioRepository.existsById(id)) {
+  			comentarioRepository.deleteById(id);	
+  		}else {
+  			throw new ComentarioNotFoundException(id);
+  		}
+  	}
+  	
+  	public Comentario getById(Long id){
+  			return comentarioRepository.findById(id)
+  					.orElseThrow(() ->new ComentarioNotFoundException(id));
+  	}
+    
+ // Este método busca un comentario por su ID y lo actualiza con los detalles proporcionados
+    public Comentario patchComentario(Long id, Comentario comentarioEdit) {
+        // Busca el comentario en el repositorio por su ID. Si no se encuentra, lanza una excepción UserNotFoundException
+        Comentario comentario = comentarioRepository.findById(id)
+            .orElseThrow(() -> new ComentarioNotFoundException(id));
+        // Actualiza el cuerpo del comentario con el cuerpo proporcionado en los comentarioDetails
+        comentario.setCuerpo(comentarioEdit.getCuerpo());
+        // Actualiza la fecha_hora del comentario a la fecha y hora actuales
+        comentario.setFecha_hora(LocalDateTime.now());
+        // Guarda el comentario actualizado en el repositorio y lo devuelve
         return comentarioRepository.save(comentario);
     }
+    
 
-    public Comentario updateComentario(Comentario comentario) {
-        return comentarioRepository.save(comentario);
-    }
-
-    public void deleteComentario(Integer id) {
-        comentarioRepository.deleteById(id);
-    }
 }
